requirements.txt
# ── Core ──────────────────────────────────────────────
python-dotenv>=1.0.0
pandas>=2.1.0,<3.1
pyarrow>=14.0.0
numpy>=1.26.0

# ── Graph Database ────────────────────────────────────
neo4j>=5.14.0

# ── Deep Learning / PyTorch Geometric ─────────────────
torch>=2.2.0,<2.11
torch-geometric>=2.6.0,<3.0
# Optional performance extensions:
# - pyg-lib
# - torch-scatter
# - torch-sparse

# ── Embeddings ────────────────────────────────────────
sentence-transformers>=2.2.0

# ── Agent Framework (LangGraph, NOT LangChain chains) ─
langgraph>=0.2.0
langchain-core>=0.3.0
langchain-openai>=0.2.0

# ── UI ────────────────────────────────────────────────
streamlit>=1.29.0
pyvis>=0.3.2
streamlit-agraph>=0.0.45

# ── Evaluation & Utilities ────────────────────────────
scikit-learn>=1.3.0
tqdm>=4.66.0
loguru>=0.7.0
